benchmarks: add CodedInputStreamBenchmark
This is heavily inspired by protobuf's own benchmark suites.

Results using jdk that uses avx2 based memset:
https://gist.github.com/zpencer/6f2495018568ecb0d73fdab87cc44b8e

The data shows array, heap BB, and direct BB are similar *except* when
there are many small bytes in which case direct BB is worse.

Results using a jdk that does not use avx2 based memset:
https://gist.github.com/zpencer/100acd4271e6b8e423dabeb1f849981e

Similar to above except direct BB is relatively slower with large byte
arrays. This is because for a heap array, #hasArray() is true, and the
code paths are the same as the array case.