java_grpc_library.bzl: workaround strict_deps
This is a sync from our internal copy, so has lots of small changes.
But the most important change is strict_deps issues are worked around
within the rule. This means there should be no longer any reason to
specify more than one dep.

Bazel 0.22 started the migration from the 'proto' provider to ProtoInfo,
which this includes. That means this is only compatible with Bazel 0.22
and later. But if we use the old-style provider it will probably stop
working with 0.23 or 0.24.

Fixes #5314