core: add Grpc.TRANSPORT_ATTR_CLIENT_EAG_ATTRS
This is needed for GRPCLB pick_first support, which needs to attach
tokens to headers, and the tokens are per server.  In pick_first, all
addresses are in a single Subchannel, thus the LoadBalancer needs to
know which backend is used for a new stream.

With this new transport attribute, the LoadBalancer can use one EAG
per backend, put the token in the EAG attributes, and then access the
token via the transport attribute available from
ClientStreamTracer.StreamInfo.