Define CreateSubchannelArgs.
As suggested by @carl-mastrangelo, using a first-class object to pass
arguments may avoid breakages when we add new arguments to
createSubchannel().  For example, a LoadBalancer may wrap Helper and
intercept createSubchannel() in a hierarchical case.  It may not be
interested in all arguments.  Passing a single CreateSubchannelArgs
will not break the parent LoadBalancer if new fields are added.

This also reduces the eventual size of Helper interface, as the
convenience createSubchannel() that accepts one EAG instead of a List
is no longer necessary. That convenience is moved into
CreateSubchannelArgs.