core: fix a bug in health check config propgation.
The condition "effectiveServiceConfig != validServiceConfig" should
have been deleted in commit 2162ad043677e3cbaac969b96fd4faa05448874b.

The condition was there before that commit because
NAME_RESOLVER_SERVICE_CONFIG was already in "attrs", thus it needed to
be re-added only if "effectiveServiceConfig" differs from the original
"validServiceConfig".

In contrast, ATTR_HEALTH_CHECKING_CONFIG is not in the original
"attrs" and always needs to be added.