Update ManualFlowControlClient example
Update the manual flow control client example to use a pattern that is applicable in all call arities (unary or streaming request).  If the call is unary request, the StreamObserver isn't returned, so the initial request can't be started from the return value, but the initial request can't be made before the call is started, so we expose a custom "request" method on the ClinetResponseObserver implementation.