`NettyClientTransport`: use `getOrCreate` for new instances of the `ChannelLogger` `AttributeKey`.
`NettyClientTransport` throws an error when the class is loaded in two separate class-loaders because of some implementation detail with the underlying `AttributeKey` class. Instead of just blowing up when a new `ChannelLogger` key can't be created, this allows multiple, isolated class-loaders in the same JVM to just get the existing key.  If this causes issues, this `AttributeKey` should be updated to use a new UUID (or something) each time it is loaded.

Improvement: Works with isolated class-loaders.
Potential Drawback: Uses the same `ChannelLogger` `AttributeKey` object across multiple, isolated classloaders.