core: add readable frame data to transport error
This change modifies the transportDataReceived handler in the HTTP/2 client
stream transport state to log the readable bytes from the frame to the error
status in the event of an unexpected EOS on a data frame. This allows users
to more easily track down the source of these unexpected stream closures.