okhttp: During testing, bind to localhost instead of 0.0.0.0
The server's address is used by the client to determine where to
connect. When binding to 0.0.0.0 the address returned by
getLocalSocketAddress() is a 0.0.0.0 address, which would seem
inappropriate for the client to use as a destination address. Somehow
this was working. However, on some systems it breaks as seen in #8080.
Using localhost is better for all systems.

Fixes #8080