Simplify Round-Robin picker logic
Since incrementAndGet is implemented the same way as updateAndGet,
except for "next = function.apply(prev)" compared to "next = prev + 1",
this change will perform the same amount of compareAndSet operations,
while also eliminating the need for bounding the index to the size using
the modulo operator.

The lambda is cheap since it does not reference any upvalues.