grpc-core: correcting a minor resource releasing issue
This change adds a try-with-resources block around readers and streams
to control the closing of those objects when the method has completed
rather than relying on the GC to deal with them.

This issue was flagged by an analysis tool via binary analysis of the
grpc-core package as part of a dependency from another project.