api: Add ClientInterceptors method overloads for ManagedChannel
The current set of ClientInterceptors methods accept a ManagedChannel,
but return a regular Channel. This effetively erases the ability
to manage the channel after these methods are used to add interceptors.

This adds 1:1 overloads of ClientInterceptors methods that both accept
and return a ManagedChannel.

This did not causes breakages interally in google3, so assumed to be
safe externally as well.