Emulate AbstractManagedChannelImplBuilder maxInboundMessageSize(int)
In dbd903c0 we started generatating AbstractManagedChannelImplBuilder
methods in a way that they are present for ABI but on recompilation only
public ManagedChannelBuilder-returning methods would be used. However,
this sorta missed maxInboundMessageSize which is now being handled by
the concrete classes (e.g, NettyChannelBuilder) instead of
ManagedChannelImplBuilder. Users on the old ABI will end up getting
ManagedChannelImplBuilder.maxInboundMessageSize() which does nothing.

So, let's just implement the method and have it jump back up to the
concrete class. Hacky, but easy and predictable and will go away once we
remove the rest of the AbstractManagedChannelImplBuilder compatibility
hacks.

Fixes #8313