Partly revert "okhttp: make okhttp dependencies compile only (#8971)"
This partly reverts b20ce178 because it is dependent on cda0e9d which is
being reverted as it is incompatible with Proguard without configuration
(e.g., isAvailable() may not be available and would need a -keep).

While the code no longer uses compileOnly, the compatibility still
remains, so it would be possible for users to exclude the okhttp
dependency if they don't use AndroidChannelBuilder.